Testing folder added
In this folder we have the integration tests written in JavaScript using the Jest testing framework, along with the Supertest library for HTTP assertions. I send HTTP request and see what the response are, the idea is to write a test that will simulate a full game being played.
For now I have implemented the registration and login of four users, one is the owner who creates the game and the other three subscribe to the game and join in different teams, the owner then starts the game. Next I will add the flow of the game with sending cards etc, I implemented tests for websocket connections.